add variable cohesion depth
